pyproject.toml
[tool.poetry]
name = "discord-mcp"
version = "0.1.0"
description = ""
authors = ["FallenDeity <61227305+FallenDeity@users.noreply.github.com>"]
readme = "README.md"
package-mode = true
packages = [
    { include = "discord_mcp", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
python-json-logger = "^3.3.0"
pytest = "^8.4.1"
click = "^8.2.1"
mcp = "^1.12.1"
discord-py = "^2.5.2"
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pre-commit = "^4.2.0"
pyright = "^1.1.403"
ruff = "^0.12.4"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportImportCycles = false
reportPrivateUsage = false
reportPrivateImportUsage = false
reportMissingModuleSource = false
reportMissingTypeStubs = false
reportUnknownMemberType = false

[tool.pytest.ini_options]
pythonpath = ["src"]
console_output_style = "progress"
